pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "requests (>=2.33.1,<3.0.0)",
    "langchain (>=1.2.14,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "mistralai (>=1.9,<2.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "datetime (>=6.0,<7.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain-mistralai (>=1.1.2,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "streamlit (>=1.56.0,<2.0.0)",
    "ragas (>=0.4.3,<0.5.0)",
    "datasets (>=4.8.4,<5.0.0)"
]

[tool.poetry]
packages = [{include = "rag", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
